Add workspace tests for IDETemplateMacros mapper
