+++ Sources/TuistCoreTesting/Graph/WorkspaceWithProjects+TestData.swift
@@ -1,9 +1,16 @@
 import Foundation
 import TSCBasic
+import TuistGraph
 @testable import TuistCore
 
 public extension WorkspaceWithProjects {
-    static func test() -> WorkspaceWithProjects {
-        WorkspaceWithProjects(workspace: .test(), projects: [.test()])
+    static func test(
+        workspace: Workspace = .test(),
+        projects: [Project] = [.test()]
+    ) -> WorkspaceWithProjects {
+        WorkspaceWithProjects(
+            workspace: workspace,
+            projects: projects
+        )
     }
 }
